+++ src/dipdup/models/__init__.py
@@ -20,7 +20,6 @@ from lru import LRU
 from pydantic.dataclasses import dataclass
 from tortoise.exceptions import OperationalError
 from tortoise.fields import relational
-from tortoise.models import MODEL
 from tortoise.models import Model as TortoiseModel
 from tortoise.queryset import BulkCreateQuery as TortoiseBulkCreateQuery
 from tortoise.queryset import BulkUpdateQuery as TortoiseBulkUpdateQuery
@@ -322,7 +321,7 @@ class BulkUpdateQuery(TortoiseBulkUpdateQuery):  # type: ignore[type-arg]
 
 
 class BulkCreateQuery(TortoiseBulkCreateQuery):  # type: ignore[type-arg]
-    async def _execute(self) -> list[MODEL]:
+    async def _execute(self) -> None:
         for model in self.objects:
             if update := ModelUpdate.from_model(
                 cast(Model, model),
@@ -330,11 +329,11 @@ class BulkCreateQuery(TortoiseBulkCreateQuery):  # type: ignore[type-arg]
             ):
                 get_pending_updates().append(update)
 
+        await super()._execute()
+
         # NOTE: A bug; raises "You should first call .save()..." otherwise
-        models: list[MODEL] = await super()._execute()
-        for model in models:
+        for model in self.objects:
             model._saved_in_db = True
-        return models
 
 
 class QuerySet(TortoiseQuerySet):  # type: ignore[type-arg]
@@ -607,7 +606,7 @@ ModelT = TypeVar('ModelT', bound=Model)
 
 
 class Schema(TortoiseModel):
-    name = fields.TextField(pk=True)
+    name = fields.TextField(primary_key=True)
     hash = fields.TextField(null=True)
     reindex = fields.EnumField(ReindexingReason, null=True)
 
@@ -619,7 +618,7 @@ class Schema(TortoiseModel):
 
 
 class Head(TortoiseModel):
-    name = fields.TextField(pk=True)
+    name = fields.TextField(primary_key=True)
     level = fields.IntField()
     hash = fields.TextField(null=True)
     timestamp = fields.DatetimeField()
@@ -632,7 +631,7 @@ class Head(TortoiseModel):
 
 
 class Index(TortoiseModel):
-    name = fields.TextField(pk=True)
+    name = fields.TextField(primary_key=True)
     type = fields.EnumField(IndexType)
     status = fields.EnumField(IndexStatus, default=IndexStatus.new)
 
@@ -657,7 +656,7 @@ class ContractKind(Enum):
 
 
 class Contract(TortoiseModel):
-    name = fields.TextField(pk=True)
+    name = fields.TextField(primary_key=True)
     address = fields.TextField(null=True)
     code_hash = fields.BigIntField(null=True)
     typename = fields.TextField(null=True)
@@ -671,7 +670,7 @@ class Contract(TortoiseModel):
 
 
 class Meta(TortoiseModel):
-    key = fields.TextField(pk=True)
+    key = fields.TextField(primary_key=True)
     value = fields.JSONField(encoder=json_dumps_plain, null=True)
 
     created_at = fields.DatetimeField(auto_now_add=True)